feat(cab-service): conformity issue and confirmity expiry date screens integration using partial view
