+++ DVSRegister/Controllers/CabController.cs
@@ -720,84 +720,6 @@ namespace DVSRegister.Controllers
             }
         }
 
-        [HttpGet("submit-certificate-information/certificate-of-confirmity-issue-date")]
-        public IActionResult ConfirmityIssueDate(bool fromSummaryPage)
-        {
-            ViewBag.fromSummaryPage = fromSummaryPage;
-            CertificateInfoSummaryViewModel summaryViewModel = GetCertificateInfoSummary();
-            DateViewModel dateViewModel = new DateViewModel();
-            if (summaryViewModel.ConformityIssueDate != null)
-            {
-                dateViewModel = GetDayMonthYear(summaryViewModel.ConformityIssueDate);
-            }
-
-            return View(dateViewModel);
-        }
-
-
-
-        /// <summary>
-        /// Updates confirmity issue date variable in session 
-        /// </summary>
-        /// <param name="viewModel"></param>
-        /// <returns></returns>
-        [HttpPost("submit-certificate-information/certificate-of-confirmity-issue-date")]
-        public IActionResult SaveConfirmityIssueDate(DateViewModel dateViewModel)
-        {
-            bool fromSummaryPage = dateViewModel.FromSummaryPage;
-            dateViewModel.FromSummaryPage = false;
-            CertificateInfoSummaryViewModel summaryViewModel = GetCertificateInfoSummary();
-            DateTime? conformityIssueDate = ValidateIssueDate(dateViewModel);
-            if (ModelState.IsValid)
-            {
-                summaryViewModel.ConformityIssueDate =conformityIssueDate;
-                HttpContext?.Session.Set("CertificateInfoSummary", summaryViewModel);
-                return fromSummaryPage ? RedirectToAction("CertificateInfoSummary") : RedirectToAction("ConfirmityExpiryDate");
-            }
-            else
-            {
-                return View("ConfirmityIssueDate", dateViewModel);
-            }
-
-        }
-
-        [HttpGet("submit-certificate-information/certificate-of-confirmity-expiry-date")]
-        public IActionResult ConfirmityExpiryDate()
-        {           
-            CertificateInfoSummaryViewModel summaryViewModel = GetCertificateInfoSummary();
-            DateViewModel dateViewModel = new DateViewModel();
-            if (summaryViewModel.ConformityExpiryDate != null)
-            {
-                dateViewModel = GetDayMonthYear(summaryViewModel.ConformityExpiryDate);
-            }
-            return View(dateViewModel);
-        }
-
-        /// <summary>
-        /// Updates confirmity issue expiry date variable in session 
-        /// </summary>
-        /// <param name="viewModel"></param>
-        /// <returns></returns>
-        [HttpPost("submit-certificate-information/certificate-of-confirmity-expiry-date")]
-        public IActionResult SaveConfirmityExpiryDate(DateViewModel dateViewModel)
-        {
-            CertificateInfoSummaryViewModel summaryViewModel = GetCertificateInfoSummary();
-            DateTime? conformityExpiryDate = ValidateExpiryDate(dateViewModel, Convert.ToDateTime(summaryViewModel.ConformityIssueDate));
-            if (ModelState.IsValid)
-            {
-                summaryViewModel.ConformityExpiryDate = conformityExpiryDate;
-                HttpContext?.Session.Set("CertificateInfoSummary", summaryViewModel);
-                return RedirectToAction("CertificateInfoSummary");
-            }
-            else
-            {
-                return View("ConfirmityExpiryDate", dateViewModel);
-
-            }
-        }
-
-
-
         /// <summary>
         /// Summary page displaying data saved in session
         /// </summary>      
@@ -965,100 +887,7 @@ namespace DVSRegister.Controllers
 
         }
 
-        private DateViewModel GetDayMonthYear(DateTime? dateTime)
-        {
-            DateViewModel dateViewModel = new DateViewModel();
-            DateTime conformityIssueDate = Convert.ToDateTime(dateTime);
-            dateViewModel.Day = conformityIssueDate.Day;
-            dateViewModel.Month = conformityIssueDate.Month;
-            dateViewModel.Year = conformityIssueDate.Year;
-            return dateViewModel;
-        }
-
-        private DateTime? ValidateIssueDate(DateViewModel dateViewModel)
-        {
-            DateTime? date = null;
-            try
-            {
-                if (dateViewModel.Day == null || dateViewModel.Month == null ||dateViewModel.Year == null)
-                {
-                    if (dateViewModel.Day == null)
-                    {
-                        ModelState.AddModelError("Day", Constants.ConformityIssueDayError);
-                    }
-                    if (dateViewModel.Month == null)
-                    {
-                        ModelState.AddModelError("Month", Constants.ConformityIssueMonthError);
-                    }
-                    if (dateViewModel.Year == null)
-                    {
-                        ModelState.AddModelError("Year", Constants.ConformityIssueYearError);
-                    }
-                }
-                else
-                {
-                    date = new DateTime(Convert.ToInt32(dateViewModel.Year), Convert.ToInt32(dateViewModel.Month), Convert.ToInt32(dateViewModel.Day));
-                    if (date>DateTime.Today)
-                    {
-                        ModelState.AddModelError("ValidDate", Constants.ConformityIssuePastDateError);
-                    }
-                }
-
-            }
-            catch (Exception)
-            {
-                ModelState.AddModelError("ValidDate", Constants.ConformityIssueDateInvalidError);
-
-            }
-            return date;
-        }
-
-        private DateTime? ValidateExpiryDate(DateViewModel dateViewModel, DateTime issueDate)
-        {
-            DateTime? date = null;
-            try
-            {
-                if (dateViewModel.Day == null || dateViewModel.Month == null ||dateViewModel.Year == null)
-                {
-                    if (dateViewModel.Day == null)
-                    {
-                        ModelState.AddModelError("Day", Constants.ConformityExpiryDayError);
-                    }
-                    if (dateViewModel.Month == null)
-                    {
-                        ModelState.AddModelError("Month", Constants.ConformityExpiryMonthError);
-                    }
-                    if (dateViewModel.Year == null)
-                    {
-                        ModelState.AddModelError("Year", Constants.ConformityExpiryYearError);
-                    }
-                }
-                else
-                {
-                    date = new DateTime(Convert.ToInt32(dateViewModel.Year), Convert.ToInt32(dateViewModel.Month), Convert.ToInt32(dateViewModel.Day));
-                    var maxExpiryDate = issueDate.AddYears(2);
-                    if (date <= DateTime.Today)
-                    {
-                        ModelState.AddModelError("ValidDate", Constants.ConformityExpiryPastDateError);
-                    }
-                    else if (date <= issueDate)
-                    {
-                        ModelState.AddModelError("ValidDate", Constants.ConformityIssueDateExpiryDateError);
-                    }
-                    else if (date >= maxExpiryDate)
-                    {
-                        ModelState.AddModelError("ValidDate", Constants.ConformityMaxExpiryDateError);
-                    }
-                }
-
-            }
-            catch (Exception)
-            {
-                ModelState.AddModelError("ValidDate", Constants.ConformityExpiryDateInvalidError);
-
-            }
-            return date;
-        }
+     
         #endregion
     }
 }
